Permission service tests for connecting / disconnecting cases & contact
