+++ hrm-domain/hrm-service/service-tests/contacts.test.ts
@@ -27,8 +27,6 @@ import {
   another2,
   broken1,
   broken2,
-  case1,
-  case2,
   contact1,
   contact2,
   conversationMedia,
@@ -39,8 +37,6 @@ import {
   workerSid,
 } from './mocks';
 import './case/caseValidation';
-import * as caseApi from '@tech-matters/hrm-core/case/caseService';
-import * as caseDb from '@tech-matters/hrm-core/case/caseDataAccess';
 import * as contactApi from '@tech-matters/hrm-core/contact/contactService';
 import * as contactDb from '@tech-matters/hrm-core/contact/contactDataAccess';
 import { mockingProxy, mockSuccessfulTwilioAuthentication } from '@tech-matters/testing';
@@ -60,7 +56,7 @@ import {
   cleanupReferrals,
   deleteContactById,
   deleteJobsByContactId,
-} from './contact/db-cleanup';
+} from './contact/dbCleanup';
 import { addConversationMediaToContact } from '@tech-matters/hrm-core/contact/contactService';
 import { NewContactRecord } from '@tech-matters/hrm-core/contact/sql/contactInsertSql';
 import supertest from 'supertest';
@@ -1042,232 +1038,6 @@ describe('/contacts route', () => {
         // // Remove records to not interfere with following tests
         await deleteCsamReportsByContactId(createdContact.id, createdContact.accountSid);
         await deleteContactById(createdContact.id, createdContact.accountSid);
-      });
-    });
-  });
-
-  describe('/contacts/:contactId/connectToCase route', () => {
-    let createdContact;
-    let createdCase;
-    let anotherCreatedCase;
-    let existingContactId;
-    let nonExistingContactId;
-    let existingCaseId;
-    let anotherExistingCaseId;
-    let nonExistingCaseId;
-
-    const byGreaterId = (a, b) => b.id - a.id;
-
-    beforeEach(async () => {
-      createdContact = await contactApi.createContact(
-        accountSid,
-        workerSid,
-        <any>contact1,
-        {
-          user: twilioUser(workerSid, []),
-          can: () => true,
-        },
-      );
-      createdCase = await caseApi.createCase(case1, accountSid, workerSid);
-      anotherCreatedCase = await caseApi.createCase(case2, accountSid, workerSid);
-      const contactToBeDeleted = await contactApi.createContact(
-        accountSid,
-        workerSid,
-        <any>contact2,
-        { user: twilioUser(workerSid, []), can: () => true },
-      );
-      const caseToBeDeleted = await caseApi.createCase(case1, accountSid, workerSid);
-
-      existingContactId = createdContact.id;
-      existingCaseId = createdCase.id;
-      anotherExistingCaseId = anotherCreatedCase.id;
-      nonExistingContactId = contactToBeDeleted.id;
-      nonExistingCaseId = caseToBeDeleted.id;
-
-      await deleteContactById(contactToBeDeleted.id, contactToBeDeleted.accountSid);
-      await caseDb.deleteById(caseToBeDeleted.id, accountSid);
-    });
-
-    afterEach(async () => {
-      await deleteContactById(createdContact.id, createdContact.accountSid);
-      await caseDb.deleteById(createdCase.id, accountSid);
-      await caseDb.deleteById(anotherCreatedCase.id, accountSid);
-    });
-
-    describe('PUT', () => {
-      const subRoute = contactId => `${route}/${contactId}/connectToCase`;
-
-      test('should return 401', async () => {
-        const response = await request.put(subRoute(existingContactId)).send({});
-
-        expect(response.status).toBe(401);
-        expect(response.body.error).toBe('Authorization failed');
-      });
-
-      test('should return 200', async () => {
-        const response = await request
-          .put(subRoute(existingContactId))
-          .set(headers)
-          .send({ caseId: existingCaseId });
-
-        expect(response.status).toBe(200);
-        expect(response.body.caseId).toBe(existingCaseId);
-
-        // Test the association
-        expect(response.body.csamReports).toHaveLength(0);
-      });
-
-      // const selectCreatedCaseAudits = () =>
-      // `SELECT * FROM "Audits" WHERE "tableName" = 'Cases' AND ("oldRecord"->>'id' = '${createdCase.id}' OR "newRecord"->>'id' = '${createdCase.id}')`;
-      const countCasesAudits = async () =>
-        parseInt(
-          (
-            await db.task(t =>
-              t.any(`SELECT COUNT(*) FROM "Audits" WHERE "tableName" = 'Cases'`),
-            )
-          )[0].count,
-          10,
-        );
-
-      const selectCasesAudits = () =>
-        db.task(t => t.any(`SELECT * FROM "Audits" WHERE "tableName" = 'Cases'`));
-
-      const countContactsAudits = async () =>
-        parseInt(
-          (
-            await db.task(t =>
-              t.any(`SELECT COUNT(*) FROM "Audits" WHERE "tableName" = 'Contacts'`),
-            )
-          )[0].count,
-          10,
-        );
-      const selectContactsAudits = () =>
-        db.task(t => t.any(`SELECT * FROM "Audits" WHERE "tableName" = 'Contacts'`));
-
-      test('should create a CaseAudit', async () => {
-        const casesAuditPreviousCount = await countCasesAudits();
-        const contactsAuditPreviousCount = await countContactsAudits();
-
-        const response = await request
-          .put(subRoute(existingContactId))
-          .set(headers)
-          .send({ caseId: existingCaseId });
-
-        expect(response.status).toBe(200);
-
-        const casesAudits = await selectCasesAudits();
-        const contactsAudits = await selectContactsAudits();
-
-        // Connecting contacts to cases updates contacts, but also touches the updatedat / updatedby fields on the case
-        expect(casesAudits).toHaveLength(casesAuditPreviousCount + 1);
-        expect(contactsAudits).toHaveLength(contactsAuditPreviousCount + 1);
-
-        const lastContactAudit = contactsAudits.sort(byGreaterId)[0];
-        const { oldRecord, newRecord } = lastContactAudit;
-
-        expect(oldRecord.caseId).toBe(null);
-        expect(newRecord.caseId).toBe(existingCaseId);
-      });
-
-      test('Idempotence on connect contact to case - generates audit', async () => {
-        const response1 = await request
-          .put(subRoute(existingContactId))
-          .set(headers)
-          .send({ caseId: existingCaseId });
-
-        expect(response1.status).toBe(200);
-
-        const casesAuditPreviousCount = await countCasesAudits();
-        const contactsAuditPreviousCount = await countContactsAudits();
-
-        // repeat above operation (should do nothing but emit an audit)
-        const response2 = await request
-          .put(subRoute(existingContactId))
-          .set(headers)
-          .send({ caseId: existingCaseId });
-
-        expect(response2.status).toBe(200);
-        expect(response2.body.caseId).toBe(existingCaseId);
-
-        const casesAuditAfterCount = await countCasesAudits();
-        const contactsAuditAfterCount = await countContactsAudits();
-
-        expect(casesAuditAfterCount).toBe(casesAuditPreviousCount + 1);
-        expect(contactsAuditAfterCount).toBe(contactsAuditPreviousCount + 1);
-      });
-
-      test('Should create audit for a Contact if caseId changes', async () => {
-        const response1 = await request
-          .put(subRoute(existingContactId))
-          .set(headers)
-          .send({ caseId: existingCaseId });
-
-        expect(response1.status).toBe(200);
-
-        const casesAuditPreviousCount = await countCasesAudits();
-        const contactsAuditPreviousCount = await countContactsAudits();
-
-        // repeat above operation (should do nothing but emit an audit)
-        const response2 = await request
-          .put(subRoute(existingContactId))
-          .set(headers)
-          .send({ caseId: anotherExistingCaseId });
-
-        expect(response2.status).toBe(200);
-
-        const casesAuditAfterCount = await countCasesAudits();
-        const contactsAuditAfterCount = await countContactsAudits();
-
-        expect(casesAuditAfterCount).toBe(casesAuditPreviousCount + 1);
-        expect(contactsAuditAfterCount).toBe(contactsAuditPreviousCount + 1);
-      });
-
-      describe('use non-existent contactId', () => {
-        test('should return 404', async () => {
-          const response = await request
-            .put(subRoute(nonExistingContactId))
-            .set(headers)
-            .send({ caseId: existingCaseId });
-
-          expect(response.status).toBe(404);
-        });
-      });
-      describe('use non-existent caseId', () => {
-        test('should return 404', async () => {
-          const response = await request
-            .put(subRoute(existingContactId))
-            .set(headers)
-            .send({ caseId: nonExistingCaseId });
-
-          expect(response.status).toBe(404);
-        });
-      });
-    });
-    describe('DELETE', () => {
-      const subRoute = contactId => `${route}/${contactId}/connectToCase`;
-      test('should return 401', async () => {
-        const response = await request.delete(subRoute(existingContactId));
-
-        expect(response.status).toBe(401);
-        expect(response.body.error).toBe('Authorization failed');
-      });
-
-      test('should return 200', async () => {
-        const response = await request.delete(subRoute(existingContactId)).set(headers);
-
-        const contact = await contactDb.getById(accountSid, response.body.id);
-
-        expect(response.status).toBe(200);
-        expect(response.body.caseId).toBe(null);
-        expect(contact.caseId).toBe(null);
-      });
-
-      test('should return 404', async () => {
-        const response = await request
-          .delete(subRoute(nonExistingContactId))
-          .set(headers);
-
-        expect(response.status).toBe(404);
       });
     });
   });
